REFACTOR the phone variable data type

+++ src/main/java/com/addressbook/entities/Contact.java
@@ -8,7 +8,7 @@ public class Contact {
 	String city;
 	String state;
 	int zip;
-	int phone;
+	long phone;
 	String email;
 
 }
